IoT/MQTT/requirements.txt
numpy==2.1.1
opencv-python==4.10.0.84
paho-mqtt==2.1.0
python-dotenv==1.0.1
